Endpoint for JA to get offline results
When a JA loads the data entry page, we need to know if they've sent in results already or not (and if so, display the results). We add a new endpoint to load the results:

GET /election/<election_id>/jurisdiction/<jurisdiction_id>/round/<round_id>/results

We also move the offline results endpoints (now there are two) to their own module. They deserve it!
